+++ Libplanet.Explorer/GraphTypes/TxResultType.cs
@@ -1,0 +1,28 @@
+using GraphQL.Types;
+
+namespace Libplanet.Explorer.GraphTypes
+{
+    public class TxResultType : ObjectGraphType<TxResult>
+    {
+        public TxResultType()
+        {
+            Field<NonNullGraphType<TxStatusType>>(
+                nameof(TxResult.TxStatus),
+                description: "The transaction status.",
+                resolve: context => context.Source.TxStatus
+            );
+
+            Field<LongGraphType>(
+                nameof(TxResult.BlockIndex),
+                description: "The block index which the target transaction executed.",
+                resolve: context => context.Source.BlockIndex
+            );
+
+            Field<StringGraphType>(
+                nameof(TxResult.BlockHash),
+                description: "The block hash which the target transaction executed.",
+                resolve: context => context.Source.BlockHash
+            );
+        }
+    }
+}